Added function to remap duplicate prompts. useful to have a better naming scheme whenever a prompt is just copied (not crossovered nor mutated).
Slight changes to the base algorithm to be able to work as the baseline alg upon tweaks to its hyperparameters. basically changed the way the best individual is being tracked.

Added a repetion penalty to the mutate and crossover functions, because sometimes repetition of a sylable was happening upon generation witht the model, specifically the unsloth version